Refactor website blocking to a separate module
